Extract IAP errors and WooWPComPlan to own files
