+++ WooCommerce/Classes/ViewModels/InAppPurchases/UpgradesViewModel.swift
@@ -3,31 +3,6 @@ import SwiftUI
 import Yosemite
 import Combine
 
-enum PrePurchaseError: Error {
-    case fetchError
-    case entitlementsError
-    case inAppPurchasesNotSupported
-    case maximumSitesUpgraded
-    case userNotAllowedToUpgrade
-}
-
-enum PurchaseUpgradeError: Error {
-    case inAppPurchaseFailed(WooWPComPlan, InAppPurchaseStore.Errors)
-    case planActivationFailed(InAppPurchaseStore.Errors)
-    case unknown
-
-    var analyticErrorDetail: Error {
-        switch self {
-        case .inAppPurchaseFailed(_, let error):
-            return error
-        case .planActivationFailed(let error):
-            return error
-        default:
-            return self
-        }
-    }
-}
-
 /// ViewModel for the Upgrades View
 /// Drives the site's available In-App Purchases plan upgrades
 ///
@@ -322,10 +297,4 @@ extension UpgradesViewModel {
     func track(_ stat: WooAnalyticsStat) {
         analytics.track(stat)
     }
-}
-
-struct WooWPComPlan {
-    let wpComPlan: WPComPlanProduct
-    let wooPlan: WooPlan
-    let hardcodedPlanDataIsValid: Bool
 }
