+++ WooCommerce/Classes/ViewModels/InAppPurchases/WooWPComPlan.swift
@@ -1,0 +1,7 @@
+import Yosemite
+
+struct WooWPComPlan {
+    let wpComPlan: WPComPlanProduct
+    let wooPlan: WooPlan
+    let hardcodedPlanDataIsValid: Bool
+}
